test: make rfHandler natively testable, cover RX dedup and TX paths
Extract the vendored RCSwitch into its own lib/RCSwitch library (same
pattern as CANDriver) so native_test can lib_ignore it and substitute a
recording shim; the real RCSwitch is interrupt-driven with static frame
state a test cannot inject. The shim mirrors the API and drives
reception/transmission through static hooks.

Also fix a real cross-platform format bug surfaced by the host build:
the RfReceived payload formatted uint64_t with %llu, which is correct on
the ESP (uint64_t is unsigned long long there) but mismatched on the
LP64 host (unsigned long); an explicit cast keeps the format and the
argument in agreement on every platform.

New suite (8 cases): a received frame published as RfReceived JSON, the
100 ms duplicate-suppression window (filtered inside, re-published after,
different data passes through), and the MQTT->RF transmit path including
the zero-data and missing-field guards.

419 -> 427 native tests pass; full release gate is clean (all 7 envs
build, including esp8266_rad which uses rfHandler).

+++ lib/rfHandler/src/rfHandler.cpp
@@ -37,7 +37,9 @@ bool RfHandler::run() {
     // Filter repeated data.
     if((lastRfData.data != actualRfData.data) || (lastRfData.bitLength != actualRfData.bitLength) || (lastRfData.protocol != actualRfData.protocol)) {
       char dataOut[dataOutBufSize] = { '\0' };
-      const int32_t dataOutSize = snprintf_P(dataOut, sizeof(dataOut), rfMessageFrame, actualRfData.data, actualRfData.bitLength, actualRfData.protocol, actualRfData.pulseLength);
+      // %llu expects unsigned long long; uint64_t is that on ESP but only unsigned long on the
+      // LP64 host, so the cast keeps the format and the argument in agreement on every platform.
+      const int32_t dataOutSize = snprintf_P(dataOut, sizeof(dataOut), rfMessageFrame, static_cast<unsigned long long>(actualRfData.data), actualRfData.bitLength, actualRfData.protocol, actualRfData.pulseLength);
       const bool dataOutValid = (dataOutSize >= 0 && dataOutSize < static_cast<int32_t>(sizeof(dataOut)));
       if(!dataOutValid) { return false; }
       if(!MqttBase::sendMessage(dataOut)) { return false; }
